rename stellar_variability module
