+++ python/desc/sims_truthcatalog/stellar_variability.py
